Update 2014-10-14-running-close-to-home.md
Hi, I've made some edits and comments. I was a little cheeky at times, so be sure not to make a mass acceptance of my edits or additions.

Nice job/reporting Tim!  

 Tim starts the article by being very observatory and presents the facts and some great opposing views, but by the end of the article -- the third last paragraph --  says what he it sounds like he was really trying to say, or argue throughout.  

Is this a data page for journalists to bounce of off, or is it a critical article to challenge the system or the politicians? Maybe it's both! 

I would encourage Tim to take a stronger position, as possible as that can be without being partisan, which is always the challenge.  It sounds a little timid at first and then ends really strong. Why not start strong too?  But that's my meager first off opinion. Basically when I was reading the first half, I kept thing: So, what? So, what? This fact ... so what? Is that intentional? Is Tim and EC just giving the facts so others can make their opinion or are you trying to influence opinion? 

Great job!